refactor(worker-proxy): organize RPC relay files
