+++ src/Functions.WorkerProxy/Rpc/FunctionRpcRelayAttachmentFailure.cs
@@ -1,7 +1,7 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-namespace Azure.Functions.WorkerProxy;
+namespace Azure.Functions.WorkerProxy.Rpc;
 
 /// <summary>
 /// Identifies why a relay stream attachment was rejected.
